// src/model/user/types.ts
export interface GetUsersParams {
  page: number;
  limit: number;
}

export interface UsersResponse {
  users: User[];
  total: number;
  skip: number;
  limit: number;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  maidenName: string;
  age: number;
  gender: 'male' | 'female' | string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;

  hair: Hair;

  ip: string;
  address: Address;
  macAddress: string;
  university: string;

  bank: Bank;
  company: Company;

  ein: string;
  ssn: string;
  userAgent: string;

  crypto: Crypto;
  role: string;
}

interface Hair {
  color: string;
  type: string;
}

interface Address {
  address: string;
  city: string;
  state: string;
  stateCode: string;
  postalCode: string;
  coordinates: Coordinates;
  country: string;
}

interface Coordinates {
  lat: number;
  lng: number;
}

interface Bank {
  cardExpire: string;
  cardNumber: string;
  cardType: string;
  currency: string;
  iban: string;
}

interface Company {
  department: string;
  name: string;
  title: string;
  address: Address;
}

interface Crypto {
  coin: string;
  wallet: string;
  network: string;
}

export interface UsersPagination {
  current: number;
  pageSize: number;
}

export type SortOrder = 'asc' | 'desc';

export interface UsersSorting {
  sortBy: keyof User | null;
  order: SortOrder | null;
}